refactor(protocol): decouple phosphor-protocol from D-Bus
phosphor-protocol defined the D-Bus wire surface for every interface in
one library, so pure-compute domain libraries had to link Qt6::DBus just
to name their own result types. Split the concerns into three layers.

phosphor-dbus (new): generic, service-agnostic D-Bus utilities — a
PhosphorDBus::Client value type bound to (connection, service, path) and
the HasDBusStreaming concept. Reusable by any Qt app; no singletons.

phosphor-protocol: now the thin PlasmaZones contract layer. WireTypes.h
split into Types.h (QtCore-only value vocabulary) and Marshalling.h
(QDBusArgument operators + registration). ClientHelpers became inline
daemon-bound wrappers over PhosphorDBus::Client via a daemonClient()
value factory, so consumer call sites are unchanged.

New PhosphorProtocol::Types target links only Qt6::Core; phosphor-snap-
engine and phosphor-placement retarget to it and no longer pull Qt6::DBus
for their type vocabulary.

Full build clean, all 184 tests pass.
