+++ kwin-effect/autotilehandler.cpp
@@ -8,7 +8,7 @@
 
 #include <PhosphorProtocol/ServiceConstants.h>
 #include <PhosphorProtocol/ClientHelpers.h>
-#include <PhosphorProtocol/WireTypes.h>
+#include <PhosphorProtocol/Marshalling.h>
 
 #include <effect/effecthandler.h>
 #include <effect/effectwindow.h>
