+++ kwin-effect/plasmazoneseffect/daemon_apply.cpp
@@ -7,7 +7,7 @@
 #include <PhosphorIdentity/WindowId.h>
 #include <PhosphorProtocol/ClientHelpers.h>
 #include <PhosphorProtocol/ServiceConstants.h>
-#include <PhosphorProtocol/WireTypes.h>
+#include <PhosphorProtocol/Marshalling.h>
 
 #include <effect/effecthandler.h>
 #include <window.h>
