+++ kwin-effect/plasmazoneseffect/drag_snap.cpp
@@ -12,7 +12,7 @@
 #include <PhosphorAnimation/ProfilePaths.h>
 #include <PhosphorProtocol/ClientHelpers.h>
 #include <PhosphorProtocol/ServiceConstants.h>
-#include <PhosphorProtocol/WireTypes.h>
+#include <PhosphorProtocol/Marshalling.h>
 
 #include <effect/effecthandler.h>
 #include <window.h>
