+++ kwin-effect/plasmazoneseffect/lifecycle.cpp
@@ -5,7 +5,7 @@
 
 #include <PhosphorProtocol/ClientHelpers.h>
 #include <PhosphorProtocol/ServiceConstants.h>
-#include <PhosphorProtocol/WireTypes.h>
+#include <PhosphorProtocol/Marshalling.h>
 
 #include <effect/effecthandler.h>
 #include <core/output.h>
