+++ kwin-effect/screenchangehandler.cpp
@@ -7,7 +7,7 @@
 
 #include <PhosphorProtocol/ServiceConstants.h>
 #include <PhosphorProtocol/ClientHelpers.h>
-#include <PhosphorProtocol/WireTypes.h>
+#include <PhosphorProtocol/Marshalling.h>
 #include <PhosphorIdentity/WindowId.h>
 
 #include <effect/effecthandler.h>
