+++ libs/phosphor-dbus/PhosphorDBusConfig.cmake.in
@@ -6,8 +6,7 @@
 include(CMakeFindDependencyMacro)
 
 find_dependency(Qt6 6.6 COMPONENTS Core DBus)
-find_dependency(PhosphorDBus)
 
-include("${CMAKE_CURRENT_LIST_DIR}/PhosphorProtocolTargets.cmake")
+include("${CMAKE_CURRENT_LIST_DIR}/PhosphorDBusTargets.cmake")
 
-check_required_components(PhosphorProtocol)
+check_required_components(PhosphorDBus)
