+++ libs/phosphor-protocol/include/PhosphorProtocol/ClientHelpers.h
@@ -6,134 +6,81 @@
 #include <PhosphorProtocol/ServiceConstants.h>
 #include <PhosphorProtocol/phosphorprotocol_export.h>
 
-#include <QLoggingCategory>
+#include <PhosphorDBus/Client.h>
 
 #include <QDBusConnection>
-#include <QDBusMessage>
 #include <QDBusPendingCall>
 #include <QDBusPendingCallWatcher>
 #include <QDBusPendingReply>
+#include <QLoggingCategory>
 #include <QObject>
 #include <QVariant>
-#include <functional>
+
+#include <utility>
 
 namespace PhosphorProtocol {
+
 PHOSPHORPROTOCOL_EXPORT const QLoggingCategory& lcPhosphorProtocol();
+
+/**
+ * @brief A PhosphorDBus::Client bound to the PlasmaZones daemon.
+ *
+ * Returned by value — `Client` is a cheap handle, so this is not a singleton
+ * and holds no shared global state. It binds the canonical
+ * `org.plasmazones` service / `/PlasmaZones` object on the session bus, with
+ * call failures logged under `lcPhosphorProtocol()`.
+ *
+ * Generic, service-agnostic D-Bus plumbing lives in PhosphorDBus; this is the
+ * thin PlasmaZones-specific binding on top.
+ */
+inline PhosphorDBus::Client daemonClient()
+{
+    return PhosphorDBus::Client(QDBusConnection::sessionBus(), Service::Name, Service::ObjectPath,
+                                &lcPhosphorProtocol());
 }
 
-namespace PhosphorProtocol::ClientHelpers {
-
 /**
- * @brief Compositor-agnostic D-Bus helper utilities
+ * @brief Daemon-bound convenience wrappers around PhosphorDBus::Client.
  *
- * These functions wrap common D-Bus call patterns used by all compositor plugins
- * when communicating with the PlasmaZones daemon. They use QDBusMessage::createMethodCall
- * (not QDBusInterface) to avoid synchronous D-Bus introspection that blocks the
- * compositor thread.
+ * These talk to the canonical PlasmaZones daemon. Code that needs to address
+ * a different service should construct its own PhosphorDBus::Client.
  */
+namespace ClientHelpers {
 
-/**
- * @brief Fire-and-forget async D-Bus call with error logging.
- *
- * @param parent     QObject parent for the watcher (must outlive the call)
- * @param interface  D-Bus interface (e.g., DBus::Interface::Autotile)
- * @param method     D-Bus method name
- * @param args       Method arguments
- * @param logContext Human-readable label for the warning log
- */
+/// Fire-and-forget async call to the daemon, with error logging.
+/// @see PhosphorDBus::Client::fireAndForget
 inline void fireAndForget(QObject* parent, const QString& interface, const QString& method, const QVariantList& args,
                           const QString& logContext = {})
 {
-    if (!parent) {
-        qCWarning(lcPhosphorProtocol) << (logContext.isEmpty() ? method : logContext)
-                                      << "fireAndForget called with null parent, ignoring";
-        return;
-    }
-    QDBusMessage msg = QDBusMessage::createMethodCall(Service::Name, Service::ObjectPath, interface, method);
-    for (const QVariant& arg : args) {
-        msg << arg;
-    }
-    QDBusPendingCall pending = QDBusConnection::sessionBus().asyncCall(msg);
-    auto* watcher = new QDBusPendingCallWatcher(pending, parent);
-    const QString ctx = logContext.isEmpty() ? method : logContext;
-    QObject::connect(watcher, &QDBusPendingCallWatcher::finished, parent, [ctx](QDBusPendingCallWatcher* w) {
-        if (w->isError()) {
-            qCWarning(lcPhosphorProtocol) << ctx << "D-Bus call failed:" << w->error().message();
-        }
-        w->deleteLater();
-    });
+    daemonClient().fireAndForget(parent, interface, method, args, logContext);
 }
 
-/**
- * @brief Send a one-way notification with no expected reply.
- *
- * Marshals via @c QDBusConnection::send rather than @c asyncCall, so no
- * watcher is allocated and no reply timeout applies. Use for genuinely
- * fire-and-forget notifications (window-opened pings, drag ticks, etc.)
- * where allocating a watcher just to ignore the reply is wasted state.
- *
- * If you need error visibility on the call, use @ref fireAndForget instead.
- *
- * @param interface  D-Bus interface (e.g., DBus::Interface::Autotile)
- * @param method     D-Bus method name
- * @param args       Method arguments
- */
+/// One-way notification to the daemon with no expected reply.
+/// @see PhosphorDBus::Client::sendOneWay
 inline void sendOneWay(const QString& interface, const QString& method, const QVariantList& args = {})
 {
-    QDBusMessage msg = QDBusMessage::createMethodCall(Service::Name, Service::ObjectPath, interface, method);
-    for (const QVariant& arg : args) {
-        msg << arg;
-    }
-    QDBusConnection::sessionBus().send(msg);
+    daemonClient().sendOneWay(interface, method, args);
 }
 
-/**
- * @brief Create an async D-Bus method call and return the pending result.
- *
- * @param interface  D-Bus interface name
- * @param method     D-Bus method name
- * @param args       Method arguments
- * @return QDBusPendingCall for attaching a watcher
- */
+/// Async method call to the daemon; caller attaches its own watcher.
+/// @see PhosphorDBus::Client::asyncCall
 inline QDBusPendingCall asyncCall(const QString& interface, const QString& method, const QVariantList& args = {})
 {
-    QDBusMessage msg = QDBusMessage::createMethodCall(Service::Name, Service::ObjectPath, interface, method);
-    for (const QVariant& arg : args) {
-        msg << arg;
-    }
-    return QDBusConnection::sessionBus().asyncCall(msg);
+    return daemonClient().asyncCall(interface, method, args);
 }
 
-/**
- * @brief Synchronous D-Bus method call bounded by `Service::SyncCallTimeoutMs`.
- *
- * For the rare paths that legitimately need a blocking reply (settings
- * editor pre-load, layout import/export). Prefer @ref asyncCall whenever
- * the caller can tolerate a callback.
- *
- * Returns the raw `QDBusMessage`; check `reply.type() == ReplyMessage`
- * before reading `reply.arguments()`. Unlike `QDBusInterface::call`, this
- * does NOT perform synchronous wire introspection — it crafts the
- * `QDBusMessage` directly.
- *
- * @param interface  D-Bus interface name
- * @param method     D-Bus method name
- * @param args       Method arguments
- * @return Reply message (type ErrorMessage on failure / timeout)
- */
+/// Blocking daemon call bounded by `Service::SyncCallTimeoutMs`.
+/// Prefer @ref asyncCall whenever the caller can tolerate a callback.
+/// @see PhosphorDBus::Client::syncCall
 inline QDBusMessage syncCall(const QString& interface, const QString& method, const QVariantList& args = {})
 {
-    QDBusMessage msg = QDBusMessage::createMethodCall(Service::Name, Service::ObjectPath, interface, method);
-    for (const QVariant& arg : args) {
-        msg << arg;
-    }
-    return QDBusConnection::sessionBus().call(msg, QDBus::Block, Service::SyncCallTimeoutMs);
+    return daemonClient().syncCall(interface, method, args, Service::SyncCallTimeoutMs);
 }
 
 /**
  * @brief Async helper for loading a single daemon setting.
  *
- * Sends getSetting(name) via raw QDBusMessage, unwraps the QDBusVariant,
+ * Sends getSetting(name) to the Settings interface, unwraps the QDBusVariant,
  * and calls onValue with the extracted QVariant.
  *
  * @param parent   QObject parent for the watcher
@@ -144,10 +91,8 @@ inline QDBusMessage syncCall(const QString& interface, const QString& method, co
 template<typename Fn>
 void loadSettingAsync(QObject* parent, const QString& name, Fn&& onValue)
 {
-    QDBusMessage msg = QDBusMessage::createMethodCall(Service::Name, Service::ObjectPath, Service::Interface::Settings,
-                                                      QStringLiteral("getSetting"));
-    msg << name;
-    auto* watcher = new QDBusPendingCallWatcher(QDBusConnection::sessionBus().asyncCall(msg), parent);
+    auto* watcher = new QDBusPendingCallWatcher(
+        daemonClient().asyncCall(Service::Interface::Settings, QStringLiteral("getSetting"), {name}), parent);
     QObject::connect(watcher, &QDBusPendingCallWatcher::finished, parent,
                      [name, onValue = std::forward<Fn>(onValue)](QDBusPendingCallWatcher* w) {
                          w->deleteLater();
@@ -166,4 +111,6 @@ void loadSettingAsync(QObject* parent, const QString& name, Fn&& onValue)
                      });
 }
 
-} // namespace PhosphorProtocol::ClientHelpers
+} // namespace ClientHelpers
+
+} // namespace PhosphorProtocol
