+++ libs/phosphor-protocol/include/PhosphorProtocol/PhosphorProtocol.h
@@ -3,6 +3,7 @@
 
 #pragma once
 
-#include <PhosphorProtocol/WireTypes.h>
-#include <PhosphorProtocol/ServiceConstants.h>
 #include <PhosphorProtocol/ClientHelpers.h>
+#include <PhosphorProtocol/Marshalling.h>
+#include <PhosphorProtocol/ServiceConstants.h>
+#include <PhosphorProtocol/Types.h>
