+++ libs/phosphor-protocol/include/PhosphorProtocol/Types.h
@@ -3,11 +3,10 @@
 
 #pragma once
 
-#include <PhosphorProtocol/phosphorprotocol_export.h>
+#include <PhosphorProtocol/phosphorprotocoltypes_export.h>
 
-#include <QDBusArgument>
-#include <QDBusMetaType>
 #include <QList>
+#include <QMetaType>
 #include <QRect>
 #include <QString>
 #include <QStringList>
@@ -36,37 +35,13 @@ enum class DragBypassReason : int {
 };
 
 /// Convert to the legacy wire-format string. Returns an empty QString for None.
-PHOSPHORPROTOCOL_EXPORT QString toWireString(DragBypassReason r);
+PHOSPHORPROTOCOLTYPES_EXPORT QString toWireString(DragBypassReason r);
 
 /// Parse from the legacy wire-format string. Unknown values map to None.
-PHOSPHORPROTOCOL_EXPORT DragBypassReason bypassReasonFromWireString(const QString& s);
+PHOSPHORPROTOCOLTYPES_EXPORT DragBypassReason bypassReasonFromWireString(const QString& s);
 
 /// QDebug streaming for logging. Prints the enum name (e.g. "AutotileScreen").
-PHOSPHORPROTOCOL_EXPORT QDebug operator<<(QDebug debug, DragBypassReason r);
-
-/**
- * @brief Compile-time check that a type has QDBusArgument streaming operators.
- *
- * Use in adaptor headers to catch missing operator<</>/>> definitions at build
- * time rather than hitting a runtime "demarshalling function failed" crash.
- *
- * @code
- *   static_assert(HasDBusStreaming<MyEntry>::value,
- *       "MyEntry needs QDBusArgument operator<< and operator>> — see dbus_types.h");
- * @endcode
- */
-template<typename T, typename = void>
-struct HasDBusStreaming : std::false_type
-{
-};
-
-template<typename T>
-struct HasDBusStreaming<T,
-                        std::void_t<decltype(std::declval<QDBusArgument&>() << std::declval<const T&>()),
-                                    decltype(std::declval<const QDBusArgument&>() >> std::declval<T&>())>>
-    : std::true_type
-{
-};
+PHOSPHORPROTOCOLTYPES_EXPORT QDebug operator<<(QDebug debug, DragBypassReason r);
 
 /// D-Bus struct for batch geometry entries: (siiiis)
 ///
@@ -106,7 +81,7 @@ struct WindowGeometryEntry
 using WindowGeometryList = QList<WindowGeometryEntry>;
 
 /// D-Bus struct for autotile tile requests: (siiiissbb)
-struct PHOSPHORPROTOCOL_EXPORT TileRequestEntry
+struct PHOSPHORPROTOCOLTYPES_EXPORT TileRequestEntry
 {
     QString windowId;
     int x = 0;
@@ -305,7 +280,7 @@ struct AlgorithmInfoEntry
 using AlgorithmInfoList = QList<AlgorithmInfoEntry>;
 
 /// D-Bus struct for bridge registration result: (sss)
-struct PHOSPHORPROTOCOL_EXPORT BridgeRegistrationResult
+struct PHOSPHORPROTOCOLTYPES_EXPORT BridgeRegistrationResult
 {
     QString apiVersion;
     QString bridgeName;
@@ -392,7 +367,7 @@ struct SwapTargetResult
 /// Single source of truth replaces the effect-side
 /// m_dragBypassedForAutotile / m_cachedZoneSelectorEnabled cache that went
 /// stale after every settings reload.
-struct PHOSPHORPROTOCOL_EXPORT DragPolicy
+struct PHOSPHORPROTOCOLTYPES_EXPORT DragPolicy
 {
     bool streamDragMoved = false; ///< effect should send dragMoved D-Bus ticks
     bool showOverlay = false; ///< daemon will show zone overlay during this drag
@@ -419,7 +394,7 @@ struct PHOSPHORPROTOCOL_EXPORT DragPolicy
 /// end. Returned from WindowDragAdaptor::endDrag. The compositor plugin
 /// executes exactly the action specified; no further decisions on the effect
 /// side. Wire: (issiiiisbba(...))  —  int + 2 strings + 4 ints + string + 2 bools + EmptyZoneList
-struct PHOSPHORPROTOCOL_EXPORT DragOutcome
+struct PHOSPHORPROTOCOLTYPES_EXPORT DragOutcome
 {
     enum Action : int {
         NoOp = 0, ///< drag produced no state change (e.g. cancelled with no prior snap)
@@ -490,82 +465,12 @@ struct PreTileGeometryEntry
 
 using PreTileGeometryList = QList<PreTileGeometryEntry>;
 
-// QDBusArgument streaming operators (implemented in wiretypes.cpp)
-PHOSPHORPROTOCOL_EXPORT QDBusArgument& operator<<(QDBusArgument& arg, const WindowGeometryEntry& e);
-PHOSPHORPROTOCOL_EXPORT const QDBusArgument& operator>>(const QDBusArgument& arg, WindowGeometryEntry& e);
-PHOSPHORPROTOCOL_EXPORT QDBusArgument& operator<<(QDBusArgument& arg, const TileRequestEntry& e);
-PHOSPHORPROTOCOL_EXPORT const QDBusArgument& operator>>(const QDBusArgument& arg, TileRequestEntry& e);
-PHOSPHORPROTOCOL_EXPORT QDBusArgument& operator<<(QDBusArgument& arg, const SnapAllResultEntry& e);
-PHOSPHORPROTOCOL_EXPORT const QDBusArgument& operator>>(const QDBusArgument& arg, SnapAllResultEntry& e);
-PHOSPHORPROTOCOL_EXPORT QDBusArgument& operator<<(QDBusArgument& arg, const SnapConfirmationEntry& e);
-PHOSPHORPROTOCOL_EXPORT const QDBusArgument& operator>>(const QDBusArgument& arg, SnapConfirmationEntry& e);
-PHOSPHORPROTOCOL_EXPORT QDBusArgument& operator<<(QDBusArgument& arg, const WindowOpenedEntry& e);
-PHOSPHORPROTOCOL_EXPORT const QDBusArgument& operator>>(const QDBusArgument& arg, WindowOpenedEntry& e);
-PHOSPHORPROTOCOL_EXPORT QDBusArgument& operator<<(QDBusArgument& arg, const WindowStateEntry& e);
-PHOSPHORPROTOCOL_EXPORT const QDBusArgument& operator>>(const QDBusArgument& arg, WindowStateEntry& e);
-PHOSPHORPROTOCOL_EXPORT QDBusArgument& operator<<(QDBusArgument& arg, const UnfloatRestoreResult& e);
-PHOSPHORPROTOCOL_EXPORT const QDBusArgument& operator>>(const QDBusArgument& arg, UnfloatRestoreResult& e);
-PHOSPHORPROTOCOL_EXPORT QDBusArgument& operator<<(QDBusArgument& arg, const ZoneGeometryRect& e);
-PHOSPHORPROTOCOL_EXPORT const QDBusArgument& operator>>(const QDBusArgument& arg, ZoneGeometryRect& e);
-PHOSPHORPROTOCOL_EXPORT QDBusArgument& operator<<(QDBusArgument& arg, const EmptyZoneEntry& e);
-PHOSPHORPROTOCOL_EXPORT const QDBusArgument& operator>>(const QDBusArgument& arg, EmptyZoneEntry& e);
-PHOSPHORPROTOCOL_EXPORT QDBusArgument& operator<<(QDBusArgument& arg, const SnapAssistCandidate& e);
-PHOSPHORPROTOCOL_EXPORT const QDBusArgument& operator>>(const QDBusArgument& arg, SnapAssistCandidate& e);
-PHOSPHORPROTOCOL_EXPORT QDBusArgument& operator<<(QDBusArgument& arg, const NamedZoneGeometry& e);
-PHOSPHORPROTOCOL_EXPORT const QDBusArgument& operator>>(const QDBusArgument& arg, NamedZoneGeometry& e);
-PHOSPHORPROTOCOL_EXPORT QDBusArgument& operator<<(QDBusArgument& arg, const AlgorithmInfoEntry& e);
-PHOSPHORPROTOCOL_EXPORT const QDBusArgument& operator>>(const QDBusArgument& arg, AlgorithmInfoEntry& e);
-PHOSPHORPROTOCOL_EXPORT QDBusArgument& operator<<(QDBusArgument& arg, const BridgeRegistrationResult& e);
-PHOSPHORPROTOCOL_EXPORT const QDBusArgument& operator>>(const QDBusArgument& arg, BridgeRegistrationResult& e);
-PHOSPHORPROTOCOL_EXPORT QDBusArgument& operator<<(QDBusArgument& arg, const MoveTargetResult& e);
-PHOSPHORPROTOCOL_EXPORT const QDBusArgument& operator>>(const QDBusArgument& arg, MoveTargetResult& e);
-PHOSPHORPROTOCOL_EXPORT QDBusArgument& operator<<(QDBusArgument& arg, const FocusTargetResult& e);
-PHOSPHORPROTOCOL_EXPORT const QDBusArgument& operator>>(const QDBusArgument& arg, FocusTargetResult& e);
-PHOSPHORPROTOCOL_EXPORT QDBusArgument& operator<<(QDBusArgument& arg, const CycleTargetResult& e);
-PHOSPHORPROTOCOL_EXPORT const QDBusArgument& operator>>(const QDBusArgument& arg, CycleTargetResult& e);
-PHOSPHORPROTOCOL_EXPORT QDBusArgument& operator<<(QDBusArgument& arg, const SwapTargetResult& e);
-PHOSPHORPROTOCOL_EXPORT const QDBusArgument& operator>>(const QDBusArgument& arg, SwapTargetResult& e);
-PHOSPHORPROTOCOL_EXPORT QDBusArgument& operator<<(QDBusArgument& arg, const RestoreTargetResult& e);
-PHOSPHORPROTOCOL_EXPORT const QDBusArgument& operator>>(const QDBusArgument& arg, RestoreTargetResult& e);
-PHOSPHORPROTOCOL_EXPORT QDBusArgument& operator<<(QDBusArgument& arg, const PreTileGeometryEntry& e);
-PHOSPHORPROTOCOL_EXPORT const QDBusArgument& operator>>(const QDBusArgument& arg, PreTileGeometryEntry& e);
-PHOSPHORPROTOCOL_EXPORT QDBusArgument& operator<<(QDBusArgument& arg, const DragPolicy& p);
-PHOSPHORPROTOCOL_EXPORT const QDBusArgument& operator>>(const QDBusArgument& arg, DragPolicy& p);
-PHOSPHORPROTOCOL_EXPORT QDBusArgument& operator<<(QDBusArgument& arg, const DragOutcome& o);
-PHOSPHORPROTOCOL_EXPORT const QDBusArgument& operator>>(const QDBusArgument& arg, DragOutcome& o);
-
-// Compile-time verification that all D-Bus struct types have streaming operators.
-// If you add a new struct above and forget the operator<</>> declarations, the
-// build will fail here with a clear message instead of crashing at runtime.
-static_assert(HasDBusStreaming<WindowGeometryEntry>::value, "WindowGeometryEntry missing QDBusArgument operators");
-static_assert(HasDBusStreaming<TileRequestEntry>::value, "TileRequestEntry missing QDBusArgument operators");
-static_assert(HasDBusStreaming<SnapAllResultEntry>::value, "SnapAllResultEntry missing QDBusArgument operators");
-static_assert(HasDBusStreaming<SnapConfirmationEntry>::value, "SnapConfirmationEntry missing QDBusArgument operators");
-static_assert(HasDBusStreaming<WindowOpenedEntry>::value, "WindowOpenedEntry missing QDBusArgument operators");
-static_assert(HasDBusStreaming<WindowStateEntry>::value, "WindowStateEntry missing QDBusArgument operators");
-static_assert(HasDBusStreaming<UnfloatRestoreResult>::value, "UnfloatRestoreResult missing QDBusArgument operators");
-static_assert(HasDBusStreaming<ZoneGeometryRect>::value, "ZoneGeometryRect missing QDBusArgument operators");
-static_assert(HasDBusStreaming<EmptyZoneEntry>::value, "EmptyZoneEntry missing QDBusArgument operators");
-static_assert(HasDBusStreaming<SnapAssistCandidate>::value, "SnapAssistCandidate missing QDBusArgument operators");
-static_assert(HasDBusStreaming<NamedZoneGeometry>::value, "NamedZoneGeometry missing QDBusArgument operators");
-static_assert(HasDBusStreaming<AlgorithmInfoEntry>::value, "AlgorithmInfoEntry missing QDBusArgument operators");
-static_assert(HasDBusStreaming<BridgeRegistrationResult>::value,
-              "BridgeRegistrationResult missing QDBusArgument operators");
-static_assert(HasDBusStreaming<MoveTargetResult>::value, "MoveTargetResult missing QDBusArgument operators");
-static_assert(HasDBusStreaming<FocusTargetResult>::value, "FocusTargetResult missing QDBusArgument operators");
-static_assert(HasDBusStreaming<CycleTargetResult>::value, "CycleTargetResult missing QDBusArgument operators");
-static_assert(HasDBusStreaming<SwapTargetResult>::value, "SwapTargetResult missing QDBusArgument operators");
-static_assert(HasDBusStreaming<RestoreTargetResult>::value, "RestoreTargetResult missing QDBusArgument operators");
-static_assert(HasDBusStreaming<PreTileGeometryEntry>::value, "PreTileGeometryEntry missing QDBusArgument operators");
-static_assert(HasDBusStreaming<DragPolicy>::value, "DragPolicy missing QDBusArgument operators");
-static_assert(HasDBusStreaming<DragOutcome>::value, "DragOutcome missing QDBusArgument operators");
-
-/// Call once at startup (daemon and plugin) to register types with Qt D-Bus
-PHOSPHORPROTOCOL_EXPORT void registerWireTypes();
-
 } // namespace PhosphorProtocol
 
-// Must be outside namespace for Qt meta-type system
+// Must be outside namespace for Qt meta-type system. Q_DECLARE_METATYPE only
+// needs <QMetaType> (QtCore) — these declarations let the value types travel
+// in a QVariant without dragging in QtDBus. The matching QDBusArgument
+// marshallers live in Marshalling.h.
 Q_DECLARE_METATYPE(PhosphorProtocol::WindowGeometryEntry)
 Q_DECLARE_METATYPE(PhosphorProtocol::WindowGeometryList)
 Q_DECLARE_METATYPE(PhosphorProtocol::TileRequestEntry)
