+++ libs/phosphor-protocol/tests/test_phosphorprotocol.cpp
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2026 fuddlesworth
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
-#include <PhosphorProtocol/WireTypes.h>
+#include <PhosphorProtocol/Marshalling.h>
 #include <PhosphorProtocol/ServiceConstants.h>
 
 #include <QTest>
