+++ src/dbus/autotileadaptor.h
@@ -5,7 +5,7 @@
 
 #include "plasmazones_export.h"
 
-#include <PhosphorProtocol/WireTypes.h>
+#include <PhosphorProtocol/Marshalling.h>
 #include <QDBusAbstractAdaptor>
 #include <QObject>
 #include <QRect>
